refactor(plugin): extract surface_store.rs (#1079)
Moves the ~13 host_ss_* SurfaceStore callbacks (including Linux-only
register_texture / register_pixel_buffer_with_timeline / lookup_texture
/ update_image_layout pairs with non-Linux platform stubs) +
HOST_SURFACE_STORE_VTABLE + host_surface_store_vtable() accessor +
the tier-1 wire-format test module out of host_services/mod.rs into
a sibling submodule.

Three GpuContextLimitedAccess callbacks that were physically co-
located in the source (acquire_pixel_buffer / get_pixel_buffer /
resolve_pixel_buffer_by_surface_id) plus the pixel_format_from_raw
helper stay in mod.rs — they belong with the GpuContext vtable and
will move to gpu_context.rs in the next commit. mod.rs re-exports
HOST_SURFACE_STORE_VTABLE and host_surface_store_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
